refactor: move dbtool to new file
