+++ bin/reth/src/lib.rs
@@ -12,6 +12,7 @@ pub mod chain;
 pub mod cli;
 pub mod db;
 pub mod dirs;
+pub mod drop_stage;
 pub mod dump_stage;
 pub mod node;
 pub mod p2p;
@@ -21,3 +22,11 @@ pub mod stage;
 pub mod test_eth_chain;
 pub mod test_vectors;
 pub mod utils;
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq, PartialOrd, Ord, clap::ValueEnum)]
+enum StageEnum {
+    Headers,
+    Bodies,
+    Senders,
+    Execution,
+}
